+++ client/src/pages/HomePage.js
@@ -1,9 +1,11 @@
+import { Navigate } from "react-router-dom";
+import axios from "axios";
+
 import AuthedLayout from "../components/AuthedLayout";
 import Dashboard from "../components/Dashboard";
 import { useAuth } from "../context/AuthContext";
-import { Navigate } from "react-router-dom";
 
-const HomePage = () => {
+export const HomePage = () => {
   const { authedUser } = useAuth();
 
   if (!authedUser) {
@@ -17,4 +19,17 @@ const HomePage = () => {
   );
 };
 
-export default HomePage;
+export const homePageLoader = async ({ request, params }) => {
+  const { data } = await axios.get("/api/session");
+  if (data) {
+    const userTeamsData = await axios.get(`/api/users/${data.id}/user-teams`);
+    const userTeams = userTeamsData.data.filter(
+      (team) => team.status !== "invited" && team.status !== "requested"
+    );
+    const invites = userTeamsData.data.filter(
+      (team) => team.status === "invited"
+    );
+    return { userTeams, invites };
+  }
+  return null;
+};
